export all utls handling in the app comp to a hook
